Factorization of the NoConnection item from HomeScreen. Simplification of the HomeViewModel Implementation of the eventual connectivity in the LikedScreen and LikedViewModel.
