+++ app/src/main/java/com/example/restau/presentation/common/NoConnection.kt
@@ -1,0 +1,49 @@
+package com.example.restau.presentation.common
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.restau.R
+import com.example.restau.ui.theme.Poppins
+
+@Composable
+fun NoConnection(
+    modifier: Modifier = Modifier
+) {
+    Column(
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = modifier.fillMaxSize().padding(5.dp)
+    ) {
+        Icon(
+            painter = painterResource(id = R.drawable.signal),
+            contentDescription = "Check your connectivity and start your experience!",
+            tint = Color.Gray,
+            modifier = Modifier.size(65.dp)
+        )
+        Spacer(modifier = Modifier.height(5.dp))
+        Text(
+            text = "Check your connectivity and start your experience!",
+            color = Color.Gray,
+            fontFamily = Poppins,
+            fontSize = 25.sp,
+            fontWeight = FontWeight.SemiBold,
+            textAlign = TextAlign.Center
+        )
+    }
+}
