+++ app/src/main/java/com/example/restau/presentation/home/HomeViewModel.kt
@@ -1,30 +1,20 @@
 package com.example.restau.presentation.home
 
 import android.app.Application
-import android.content.Context
-import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.restau.domain.model.Restaurant
 import com.example.restau.domain.model.User
 import com.example.restau.domain.usecases.analyticsUseCases.AnalyticsUseCases
 import com.example.restau.domain.usecases.restaurantUseCases.RestaurantUseCases
 import com.example.restau.domain.usecases.userUseCases.UserUseCases
-import com.example.restau.utils.NetworkConnectionState
-import com.example.restau.utils.currentConnectivityState
 import com.example.restau.utils.getConnectivityAsStateFlow
-import com.example.restau.utils.observeConnectivityAsFlow
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import java.util.Date
 import java.util.concurrent.TimeUnit
@@ -38,11 +28,7 @@ class HomeViewModel @Inject constructor(
     private val application: Application
 ): AndroidViewModel(application) {
 
-    private val context: Context
-        get() = getApplication<Application>().applicationContext
-
-    // Expose the network state as a StateFlow
-    val isConnected: StateFlow<Boolean> = context.getConnectivityAsStateFlow(viewModelScope)
+    val isConnected: StateFlow<Boolean> = application.getConnectivityAsStateFlow(viewModelScope)
 
     var state by mutableStateOf(HomeState())
         private set
