+++ app/src/main/java/com/example/restau/presentation/liked/LikedScreen.kt
@@ -14,6 +14,8 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -29,6 +31,7 @@ import androidx.navigation.NavHostController
 import com.example.restau.R
 import com.example.restau.presentation.common.DynamicTopBar
 import com.example.restau.presentation.common.LoadingCircle
+import com.example.restau.presentation.common.NoConnection
 import com.example.restau.presentation.common.RestaurantsLazyList
 import com.example.restau.presentation.common.TopBarAction
 import com.example.restau.presentation.navigation.Route
@@ -42,7 +45,9 @@ fun LikedScreen(
 
     val user = likedViewModel.currentUser
 
-    LaunchedEffect(Unit) {
+    val isConnected by likedViewModel.isConnected.collectAsState()
+
+    LaunchedEffect(isConnected) {
         likedViewModel.onEvent(LikedEvent.ScreenLaunched)
     }
 
@@ -65,6 +70,7 @@ fun LikedScreen(
         }
     ) {
         LikedContent(
+            showFallback = likedViewModel.showFallback,
             state = likedViewModel.state,
             onLike = {documentId, delete ->
                 likedViewModel.onEvent(LikedEvent.SendLike(documentId, delete))
@@ -84,26 +90,31 @@ fun LikedContent(
     state: LikedState,
     onLike: (String, Boolean) -> Unit,
     modifier: Modifier = Modifier,
-    navController: NavHostController
+    navController: NavHostController,
+    showFallback: Boolean
 ) {
 
     Column(
         modifier = modifier.fillMaxSize()
     ) {
-        if (state.restaurants.isEmpty()) {
-            LoadingCircle()
-        } else if (state.isLiked.all { !it }) {
-            NoLikedRestaurants()
+        if (!showFallback) {
+            if (state.restaurants.isEmpty()) {
+                LoadingCircle()
+            } else if (state.isLiked.all { !it }) {
+                NoLikedRestaurants()
+            } else {
+                Spacer(modifier = Modifier.height(22.dp))
+                RestaurantsLazyList(
+                    restaurants = state.restaurants,
+                    isNew = state.isNew,
+                    isShown = state.isLiked,
+                    isLiked = state.isLiked,
+                    onLike = onLike,
+                    onClick = { navController.navigate(Route.RestaurantScreen.route + it) }
+                )
+            }
         } else {
-            Spacer(modifier = Modifier.height(22.dp))
-            RestaurantsLazyList(
-                restaurants = state.restaurants,
-                isNew = state.isNew,
-                isShown = state.isLiked,
-                isLiked = state.isLiked,
-                onLike = onLike,
-                onClick = { navController.navigate(Route.RestaurantScreen.route + it) }
-            )
+            NoConnection()
         }
     }
 }
