+++ app/src/main/java/com/example/restau/presentation/liked/LikedViewModel.kt
@@ -1,8 +1,11 @@
 package com.example.restau.presentation.liked
 
+import android.app.Application
+import android.content.Context
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
+import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.restau.domain.model.Restaurant
@@ -10,8 +13,10 @@ import com.example.restau.domain.model.User
 import com.example.restau.domain.usecases.analyticsUseCases.AnalyticsUseCases
 import com.example.restau.domain.usecases.restaurantUseCases.RestaurantUseCases
 import com.example.restau.domain.usecases.userUseCases.UserUseCases
+import com.example.restau.utils.getConnectivityAsStateFlow
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import java.util.Date
 import java.util.concurrent.TimeUnit
@@ -22,7 +27,16 @@ class LikedViewModel @Inject constructor(
     private val analyticsUseCases: AnalyticsUseCases,
     private val userUseCases: UserUseCases,
     private val restaurantUseCases: RestaurantUseCases,
-): ViewModel() {
+    private val application: Application
+): AndroidViewModel(application) {
+
+    private val context: Context
+        get() = getApplication<Application>().applicationContext
+
+    val isConnected: StateFlow<Boolean> = context.getConnectivityAsStateFlow(viewModelScope)
+
+    var showFallback by mutableStateOf(false)
+        private set
 
     var state by mutableStateOf(LikedState())
         private set
@@ -36,7 +50,14 @@ class LikedViewModel @Inject constructor(
         viewModelScope.launch(Dispatchers.IO) {
             val user = userUseCases.getUserObject()
             currentUser = user
-            getRestaurants()
+            val restaurants = restaurantUseCases.getRestaurants()
+            if (!isConnected.value && restaurants.isEmpty()) {
+                showFallback = true
+            } else {
+                showFallback = false
+                getRestaurants()
+            }
+
         }
     }
